refactor(sidebar): colocate Zellij pid backfill with pane metrics
Move the existing functions unchanged; the mux resolver remains the backend
seam below metrics.
